update csharp templates to be more templatey
