+++ src/ai/.x/templates/openai-chat-cs/Program.cs
@@ -1,29 +1,1 @@
-//
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE.md file in the project root for full license information.
-//
-
-using System;
-
-public class Program
-{
-    public static void Main(string[] args)
-    {
-        var openAIAPIKey = Environment.GetEnvironmentVariable("AZURE_OPENAI_API_KEY") ?? "{AZURE_OPENAI_API_KEY}";
-        var openAIEndpoint = Environment.GetEnvironmentVariable("AZURE_OPENAI_ENDPOINT") ?? "{AZURE_OPENAI_ENDPOINT}";
-        var openAIChatDeploymentName = Environment.GetEnvironmentVariable("AZURE_OPENAI_CHAT_DEPLOYMENT") ?? "{AZURE_OPENAI_CHAT_DEPLOYMENT}";
-        var openAISystemPrompt = Environment.GetEnvironmentVariable("AZURE_OPENAI_SYSTEM_PROMPT") ?? "{AZURE_OPENAI_SYSTEM_PROMPT}";
-
-		var chat = new {ClassName}(openAIEndpoint, openAIAPIKey, openAIChatDeploymentName, openAISystemPrompt);
-
-        while (true)
-        {
-            Console.Write("User: ");
-            var userPrompt = Console.ReadLine();
-            if (string.IsNullOrEmpty(userPrompt) || userPrompt == "exit") break;
-
-            var response = chat.GetChatCompletion(userPrompt);
-            Console.WriteLine($"\nAssistant: {response}\n");
-        }
-    }
-}
+{{@include openai-cs/Program.cs}}
