+++ src/ai/.x/templates/openai-chat-streaming-cs/OpenAIChatCompletionsStreaming.csproj._
@@ -1,16 +1,1 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <EnableDefaultCompileItems>true</EnableDefaultCompileItems>
-    <OutputType>Exe</OutputType>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Azure.Identity" Version="1.12.0" />
-    <PackageReference Include="Azure.AI.OpenAI" Version="2.0.0-beta.5" />
-  </ItemGroup>
-
-</Project>
+{{@include openai-cs/openai-cs.csproj._}}
